+++ Core.lua
@@ -88,11 +88,16 @@ function MythicPlusUtility:RefreshConfig()
     if self.Frame then self.Frame:ProfileChange() end
 end
 
-function MythicPlusUtility:onTalentFrameShow() MythicPlusUtility.TalentFrameHighlight:UpdateAnchers() end
+function MythicPlusUtility:onTalentFrameShow()
+    if MythicPlusUtility.Frame and MythicPlusUtility.Frame:IsShown() then
+        MythicPlusUtility.TalentFrameHighlight:UpdateAnchers()
+        MythicPlusUtility.TalentFrameHighlight:ShowRelevant()
+    end
+end
 
 function MythicPlusUtility:onUtilityWindowSetShown()
-    if MythicPlusUtility.Frame:IsVisible() then
-        MythicPlusUtility.TalentFrameHighlight:HideAll()
+    if MythicPlusUtility.Frame:IsShown() then
+        MythicPlusUtility.TalentFrameHighlight:UpdateAnchers()
         MythicPlusUtility.TalentFrameHighlight:ShowRelevant()
     else
         MythicPlusUtility.TalentFrameHighlight:HideAll()
@@ -103,6 +108,7 @@ function MythicPlusUtility:onUtilityWindowChangeInstance()
     if MythicPlusUtility.Frame then
         MythicPlusUtility.TalentFrameHighlight:UpdateAnchers()
         MythicPlusUtility.TalentFrameHighlight:UpdateHighlight()
+        if MythicPlusUtility.Frame:IsShown() then MythicPlusUtility.TalentFrameHighlight:ShowRelevant() end
     end
 end
 
@@ -119,19 +125,20 @@ end
 
 function MythicPlusUtility:ACTIVE_PLAYER_SPECIALIZATION_CHANGED(event)
     self.db.char.currentSpec = C_SpecializationInfo.GetSpecializationInfo(C_SpecializationInfo.GetSpecialization())
-    if self.Frame and self.Frame:IsVisible() then
+    if self.Frame and self.Frame:IsShown() then
         self:CreateCurrentAbilitiesList()
         self.Frame:ChangeInstance()
-        C_Timer.NewTimer(0.5, function() MythicPlusUtility.TalentFrameHighlight.UpdateAnchers(MythicPlusUtility) end)
+       -- C_Timer.NewTimer(0.5, function() MythicPlusUtility.TalentFrameHighlight:UpdateAnchers() end)
+       C_Timer.NewTimer(0.5, function() MythicPlusUtility:onUtilityWindowChangeInstance() end)
     else
         self.db.char.changedSpec = true
     end
 
-    self:onUtilityWindowChangeInstance()
+    --self:onUtilityWindowChangeInstance()
 end
 
 function MythicPlusUtility:TRAIT_CONFIG_UPDATED(event)
-    if self.Frame and self.Frame:IsVisible() then
+    if self.Frame and self.Frame:IsShown() then
         self:UpdateCurrentAbilitiesList()
         self.Frame:ChangeInstance()
     else
@@ -149,7 +156,7 @@ function MythicPlusUtility:PLAYER_ENTERING_WORLD(event, isInitialLogin, isReload
                 MythicPlusUtility.Frame:SetShownHandler(true)
                 MythicPlusUtility.Frame:ChangeInstance()
             else
-                if MythicPlusUtility.Frame and MythicPlusUtility.Frame:IsVisible() then
+                if MythicPlusUtility.Frame and MythicPlusUtility.Frame:IsShown() then
                     MythicPlusUtility.Frame:SetShownHandler(false)
                 end
             end
@@ -158,11 +165,11 @@ function MythicPlusUtility:PLAYER_ENTERING_WORLD(event, isInitialLogin, isReload
 end
 
 function MythicPlusUtility:CHALLENGE_MODE_START(event)
-    if self.Frame and self.db.profile.hideOnStart and self.Frame:IsVisible() then self.Frame:SetShownHandler(false) end
+    if self.Frame and self.db.profile.hideOnStart and self.Frame:IsShown() then self.Frame:SetShownHandler(false) end
 end
 
 function MythicPlusUtility:UPDATE_VEHICLE_ACTIONBAR(event)
-    if self.Frame and self.Frame:IsVisible() then
+    if self.Frame and self.Frame:IsShown() then
         self:UpdateCurrentAbilitiesList(true)
         self.Frame:ChangeInstance()
     else
@@ -248,7 +255,7 @@ function MythicPlusUtility:ToggleAbilitiesFrame()
         self.Frame:SetShownHandler(true)
         return
     end
-    self.Frame:SetShownHandler(not self.Frame:IsVisible())
+    self.Frame:SetShownHandler(not self.Frame:IsShown())
 end
 
 function MythicPlusUtility:OpenSettings(inCombat, frame) if not inCombat then Settings.OpenToCategory(frame.name) end end